Now using WebClient to download XML file instead of HttpClient.
